Setup data and created displaySchedules function
I've created arrays that store all the data which will be used throughout the program.
This incluces train names, schedules, number of stops, departure times, and prices.
Also setup a basic output function that shows schedule of all the trains however
the formatting of the output will probably be changed